% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-network.R
\name{as_sonnet_network}
\alias{as_sonnet_network}
\title{Create sonnet_network S3 class wrapper}
\usage{
as_sonnet_network(network)
}
\arguments{
\item{network}{SonnetNetwork R6 object.}
}
\value{
Object with sonnet_network class.
}
\description{
Create sonnet_network S3 class wrapper
}
\keyword{internal}
