% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-parse.R
\name{create_edges_df}
\alias{create_edges_df}
\title{Create Edge Data Frame}
\usage{
create_edges_df(from, to, weight = NULL, directed = FALSE)
}
\arguments{
\item{from}{Vector of source node indices.}

\item{to}{Vector of target node indices.}

\item{weight}{Vector of edge weights.}

\item{directed}{Logical. Is the network directed?}
}
\value{
Data frame with edge information.
}
\description{
Create Edge Data Frame
}
\keyword{internal}
