% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-grid.R
\name{create_grid_grob}
\alias{create_grid_grob}
\title{Create Grid Grob Tree}
\usage{
create_grid_grob(network, title = NULL)
}
\arguments{
\item{network}{A sonnet_network object.}

\item{title}{Optional plot title.}
}
\value{
A grid gTree object.
}
\description{
Create a complete grid grob tree for the network (without drawing).
}
\keyword{internal}
