% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-special.R
\name{draw_cross}
\alias{draw_cross}
\title{Draw Cross/Plus Node}
\usage{
draw_cross(
  x,
  y,
  size,
  fill,
  border_color,
  border_width,
  alpha = 1,
  thickness = 0.3,
  ...
)
}
\description{
Draw Cross/Plus Node
}
\keyword{internal}
