% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-special.R
\name{draw_ellipse}
\alias{draw_ellipse}
\title{Draw Ellipse Node}
\usage{
draw_ellipse(
  x,
  y,
  size,
  fill,
  border_color,
  border_width,
  alpha = 1,
  aspect = 0.6,
  ...
)
}
\description{
Draw Ellipse Node
}
\keyword{internal}
