% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-basic.R
\name{draw_pentagon}
\alias{draw_pentagon}
\title{Draw Pentagon Node}
\usage{
draw_pentagon(x, y, size, fill, border_color, border_width, alpha = 1, ...)
}
\description{
Draw Pentagon Node
}
\keyword{internal}
