% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-special.R
\name{draw_star}
\alias{draw_star}
\title{Draw Star Node}
\usage{
draw_star(
  x,
  y,
  size,
  fill,
  border_color,
  border_width,
  alpha = 1,
  n_points = 5,
  inner_ratio = 0.4,
  ...
)
}
\description{
Draw Star Node
}
\keyword{internal}
