% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-basic.R
\name{draw_triangle}
\alias{draw_triangle}
\title{Draw Triangle Node}
\usage{
draw_triangle(x, y, size, fill, border_color, border_width, alpha = 1, ...)
}
\description{
Draw Triangle Node
}
\keyword{internal}
