% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-globals.R
\name{get_layout}
\alias{get_layout}
\title{Get a Registered Layout}
\usage{
get_layout(name)
}
\arguments{
\item{name}{Character. Name of the layout.}
}
\value{
The layout function, or NULL if not found.
}
\description{
Get a Registered Layout
}
