% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-globals.R
\name{get_shape}
\alias{get_shape}
\title{Get a Registered Shape}
\usage{
get_shape(name)
}
\arguments{
\item{name}{Character. Name of the shape.}
}
\value{
The shape drawing function, or NULL if not found.
}
\description{
Get a Registered Shape
}
