% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-globals.R
\name{get_theme}
\alias{get_theme}
\title{Get a Registered Theme}
\usage{
get_theme(name)
}
\arguments{
\item{name}{Character. Name of the theme.}
}
\value{
The theme object, or NULL if not found.
}
\description{
Get a Registered Theme
}
