% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-network.R
\name{is_sonnet_network}
\alias{is_sonnet_network}
\title{Check if object is a SonnetNetwork}
\usage{
is_sonnet_network(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical.
}
\description{
Check if object is a SonnetNetwork
}
\keyword{internal}
