% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-theme.R
\name{is_sonnet_theme}
\alias{is_sonnet_theme}
\title{Check if object is a SonnetTheme}
\usage{
is_sonnet_theme(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical.
}
\description{
Check if object is a SonnetTheme
}
\keyword{internal}
