% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-parse.R
\name{is_symmetric_matrix}
\alias{is_symmetric_matrix}
\title{Detect if Matrix is Symmetric}
\usage{
is_symmetric_matrix(m, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{m}{A matrix.}

\item{tol}{Tolerance for comparison.}
}
\value{
Logical.
}
\description{
Detect if Matrix is Symmetric
}
\keyword{internal}
