% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-geometry.R
\name{offset_point}
\alias{offset_point}
\title{Offset Point from Center}
\usage{
offset_point(x, y, toward_x, toward_y, offset)
}
\arguments{
\item{x, y}{Original point.}

\item{toward_x, toward_y}{Point to offset toward.}

\item{offset}{Distance to offset.}
}
\value{
List with x, y coordinates.
}
\description{
Calculate a point offset from another point by a given distance.
}
\keyword{internal}
