% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_blues}
\alias{palette_blues}
\title{Blues Palette}
\usage{
palette_blues(n, alpha = 1)
}
\arguments{
\item{n}{Number of colors to generate.}

\item{alpha}{Transparency (0-1).}
}
\value{
Character vector of colors.
}
\description{
Generate a blue sequential palette.
}
