% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_colorblind}
\alias{palette_colorblind}
\title{Colorblind-friendly Palette}
\usage{
palette_colorblind(n, alpha = 1)
}
\arguments{
\item{n}{Number of colors to generate.}

\item{alpha}{Transparency (0-1).}
}
\value{
Character vector of colors.
}
\description{
Generate a colorblind-friendly palette using Wong's colors.
}
\examples{
palette_colorblind(5)
}
