% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_diverging}
\alias{palette_diverging}
\title{Diverging Palette}
\usage{
palette_diverging(n, alpha = 1, midpoint = "white")
}
\arguments{
\item{n}{Number of colors to generate.}

\item{alpha}{Transparency (0-1).}

\item{midpoint}{Color for midpoint.}
}
\value{
Character vector of colors.
}
\description{
Generate a diverging color palette (blue-white-red).
}
