% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_pastel}
\alias{palette_pastel}
\title{Pastel Palette}
\usage{
palette_pastel(n, alpha = 1)
}
\arguments{
\item{n}{Number of colors to generate.}

\item{alpha}{Transparency (0-1).}
}
\value{
Character vector of colors.
}
\description{
Generate a soft pastel color palette.
}
\examples{
palette_pastel(5)
}
