% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_rainbow}
\alias{palette_rainbow}
\title{Rainbow Palette}
\usage{
palette_rainbow(n, alpha = 1)
}
\arguments{
\item{n}{Number of colors to generate.}

\item{alpha}{Transparency (0-1).}
}
\value{
Character vector of colors.
}
\description{
Generate a rainbow color palette.
}
\examples{
palette_rainbow(5)
}
