% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_reds}
\alias{palette_reds}
\title{Reds Palette}
\usage{
palette_reds(n, alpha = 1)
}
\arguments{
\item{n}{Number of colors to generate.}

\item{alpha}{Transparency (0-1).}
}
\value{
Character vector of colors.
}
\description{
Generate a red sequential palette.
}
