% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_viridis}
\alias{palette_viridis}
\title{Viridis Palette}
\usage{
palette_viridis(n, alpha = 1, option = "viridis")
}
\arguments{
\item{n}{Number of colors to generate.}

\item{alpha}{Transparency (0-1).}

\item{option}{Viridis option: "viridis", "magma", "plasma", "inferno", "cividis".}
}
\value{
Character vector of colors.
}
\description{
Generate colors from the viridis palette.
}
\examples{
palette_viridis(5)
}
