% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot.sonnet_network}
\alias{plot.sonnet_network}
\title{Plot sonnet_network Object}
\usage{
\method{plot}{sonnet_network}(x, ...)
}
\arguments{
\item{x}{A sonnet_network object.}

\item{...}{Additional arguments passed to sn_render.}
}
\value{
Invisible x.
}
\description{
Plot sonnet_network Object
}
\examples{
adj <- matrix(c(0, 1, 1, 1, 0, 1, 1, 1, 0), nrow = 3)
net <- sonnet(adj)
plot(net)
}
