% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-geometry.R
\name{point_angle}
\alias{point_angle}
\title{Calculate Angle Between Two Points}
\usage{
point_angle(x1, y1, x2, y2)
}
\arguments{
\item{x1, y1}{Start point coordinates.}

\item{x2, y2}{End point coordinates.}
}
\value{
Angle in radians.
}
\description{
Calculate Angle Between Two Points
}
\keyword{internal}
