% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-geometry.R
\name{point_distance}
\alias{point_distance}
\title{Calculate Distance Between Two Points}
\usage{
point_distance(x1, y1, x2, y2)
}
\arguments{
\item{x1, y1}{First point coordinates.}

\item{x2, y2}{Second point coordinates.}
}
\value{
Euclidean distance.
}
\description{
Calculate Distance Between Two Points
}
\keyword{internal}
