% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-geometry.R
\name{point_on_circle}
\alias{point_on_circle}
\title{Calculate Point on Circle}
\usage{
point_on_circle(cx, cy, r, angle)
}
\arguments{
\item{cx, cy}{Center coordinates.}

\item{r}{Radius.}

\item{angle}{Angle in radians.}
}
\value{
List with x, y coordinates.
}
\description{
Calculate Point on Circle
}
\keyword{internal}
