% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-print.R
\name{print.sonnet_network}
\alias{print.sonnet_network}
\title{Print sonnet_network Object}
\usage{
\method{print}{sonnet_network}(x, ...)
}
\arguments{
\item{x}{A sonnet_network object.}

\item{...}{Ignored.}
}
\value{
Invisible x.
}
\description{
Print sonnet_network Object
}
