% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-edges.R
\name{render_edge_labels_grid}
\alias{render_edge_labels_grid}
\title{Render Edge Labels}
\usage{
render_edge_labels_grid(network)
}
\arguments{
\item{network}{A SonnetNetwork object.}
}
\value{
A grid gList of label grobs.
}
\description{
Create grid grobs for edge labels.
}
\keyword{internal}
