% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-edges.R
\name{render_edges_grid}
\alias{render_edges_grid}
\title{Render All Edges}
\usage{
render_edges_grid(network)
}
\arguments{
\item{network}{A SonnetNetwork object.}
}
\value{
A grid gList of edge grobs.
}
\description{
Create grid grobs for all edges in the network.
}
\keyword{internal}
