% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-nodes.R
\name{render_node_labels_grid}
\alias{render_node_labels_grid}
\title{Render Node Labels}
\usage{
render_node_labels_grid(network)
}
\arguments{
\item{network}{A SonnetNetwork object.}
}
\value{
A grid gList of label grobs.
}
\description{
Create grid grobs for node labels.
}
\keyword{internal}
