% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-nodes.R
\name{render_nodes_grid}
\alias{render_nodes_grid}
\title{Render All Nodes}
\usage{
render_nodes_grid(network)
}
\arguments{
\item{network}{A SonnetNetwork object.}
}
\value{
A grid gList of node grobs.
}
\description{
Create grid grobs for all nodes in the network.
}
\keyword{internal}
