% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot.R
\name{render_nodes_splot}
\alias{render_nodes_splot}
\title{Render Nodes for splot}
\usage{
render_nodes_splot(
  layout,
  node_size,
  node_size2,
  node_shape,
  node_fill,
  node_border_color,
  node_border_width,
  pie_values,
  pie_colors,
  pie_border_width,
  donut_values,
  donut_colors,
  donut_border_width,
  donut_inner_ratio,
  donut_bg_color,
  donut_show_value,
  donut_value_size,
  donut_value_color,
  donut2_values,
  donut2_colors,
  donut2_inner_ratio,
  labels,
  label_size,
  label_color,
  label_position,
  usePCH = FALSE
)
}
\description{
Render Nodes for splot
}
\keyword{internal}
