% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot-labels.R
\name{resolve_stars}
\alias{resolve_stars}
\title{Resolve Stars from Various Inputs}
\usage{
resolve_stars(stars_input, p_values = NULL, n)
}
\arguments{
\item{stars_input}{User input: character vector, logical, or numeric p-values.}

\item{p_values}{P-values for computing stars if stars_input is TRUE/numeric.}

\item{n}{Number of edges.}
}
\value{
Character vector of stars.
}
\description{
Resolves significance stars from character vectors, logical, or p-values.
}
\keyword{internal}
