% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes-scales.R
\name{scale_alpha}
\alias{scale_alpha}
\title{Create an Alpha Scale}
\usage{
scale_alpha(values, range = c(0.3, 1))
}
\arguments{
\item{values}{Values to map.}

\item{range}{Output alpha range.}
}
\value{
Scaled values.
}
\description{
Map values to transparency.
}
\keyword{internal}
