% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes-scales.R
\name{scale_color_discrete}
\alias{scale_color_discrete}
\title{Create a Categorical Color Scale}
\usage{
scale_color_discrete(values, palette = "colorblind")
}
\arguments{
\item{values}{Categorical values.}

\item{palette}{Color palette.}
}
\value{
Character vector of colors.
}
\description{
Map categorical values to colors.
}
\keyword{internal}
