% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes-nodes.R
\name{scale_node_sizes}
\alias{scale_node_sizes}
\title{Scale Node Sizes}
\usage{
scale_node_sizes(values, range = c(0.03, 0.1))
}
\arguments{
\item{values}{Numeric values to scale.}

\item{range}{Target size range (min, max).}
}
\value{
Scaled size values.
}
\description{
Scale node sizes based on a numeric variable.
}
\keyword{internal}
