% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes-scales.R
\name{scale_size}
\alias{scale_size}
\title{Create a Size Scale}
\usage{
scale_size(values, range = c(0.03, 0.1), trans = "linear")
}
\arguments{
\item{values}{Values to map.}

\item{range}{Output size range.}

\item{trans}{Transformation: "linear", "sqrt", "log".}
}
\value{
Scaled values.
}
\description{
Map values to sizes.
}
\keyword{internal}
