% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes-scales.R
\name{scale_width}
\alias{scale_width}
\title{Create a Width Scale}
\usage{
scale_width(values, range = c(0.5, 3))
}
\arguments{
\item{values}{Values to map.}

\item{range}{Output width range.}
}
\value{
Scaled values.
}
\description{
Map values to line widths.
}
\keyword{internal}
