% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{summary.sonnet_network}
\alias{summary.sonnet_network}
\title{Summary of sonnet_network Object}
\usage{
\method{summary}{sonnet_network}(object, ...)
}
\arguments{
\item{object}{A sonnet_network object.}

\item{...}{Ignored.}
}
\value{
A list with network summary information (invisibly).
}
\description{
Summary of sonnet_network Object
}
