% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-builtin.R
\name{theme_sonnet_classic}
\alias{theme_sonnet_classic}
\title{Classic Theme}
\usage{
theme_sonnet_classic()
}
\value{
A SonnetTheme object.
}
\description{
Traditional network visualization style with blue nodes and gray edges.
}
\examples{
theme <- theme_sonnet_classic()
}
