% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-builtin.R
\name{theme_sonnet_colorblind}
\alias{theme_sonnet_colorblind}
\title{Colorblind-friendly Theme}
\usage{
theme_sonnet_colorblind()
}
\value{
A SonnetTheme object.
}
\description{
Theme using colors distinguishable by people with color vision deficiency.
}
\examples{
theme <- theme_sonnet_colorblind()
}
