% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-builtin.R
\name{theme_sonnet_dark}
\alias{theme_sonnet_dark}
\title{Dark Theme}
\usage{
theme_sonnet_dark()
}
\value{
A SonnetTheme object.
}
\description{
Dark background theme for presentations.
}
\examples{
theme <- theme_sonnet_dark()
}
