% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-builtin.R
\name{theme_sonnet_gray}
\alias{theme_sonnet_gray}
\title{Grayscale Theme}
\usage{
theme_sonnet_gray()
}
\value{
A SonnetTheme object.
}
\description{
Black and white theme suitable for print.
}
\examples{
theme <- theme_sonnet_gray()
}
