% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-builtin.R
\name{theme_sonnet_minimal}
\alias{theme_sonnet_minimal}
\title{Minimal Theme}
\usage{
theme_sonnet_minimal()
}
\value{
A SonnetTheme object.
}
\description{
Clean, minimal style with thin borders.
}
\examples{
theme <- theme_sonnet_minimal()
}
