% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_choice}
\alias{validate_choice}
\title{Validate Choice}
\usage{
validate_choice(x, choices, arg_name = "value")
}
\arguments{
\item{x}{Value to validate.}

\item{choices}{Allowed values.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Validate Choice
}
\keyword{internal}
