% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_color}
\alias{validate_color}
\title{Validate Color}
\usage{
validate_color(x, arg_name = "color")
}
\arguments{
\item{x}{Color to validate.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Validate Color
}
\keyword{internal}
