% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_length}
\alias{validate_length}
\title{Validate Length Match}
\usage{
validate_length(x, expected_length, arg_name = "value", allow_single = TRUE)
}
\arguments{
\item{x}{Vector to validate.}

\item{expected_length}{Expected length.}

\item{arg_name}{Argument name for error messages.}

\item{allow_single}{Allow single value (will be recycled).}
}
\description{
Validate Length Match
}
\keyword{internal}
