% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_network}
\alias{validate_network}
\title{Validate Network Object}
\usage{
validate_network(x, arg_name = "network")
}
\arguments{
\item{x}{Object to validate.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Validate Network Object
}
\keyword{internal}
