% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_range}
\alias{validate_range}
\title{Validate Numeric Range}
\usage{
validate_range(x, min = -Inf, max = Inf, arg_name = "value")
}
\arguments{
\item{x}{Value to validate.}

\item{min}{Minimum allowed value.}

\item{max}{Maximum allowed value.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Validate Numeric Range
}
\keyword{internal}
