% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes-special.R
\name{draw_pie}
\alias{draw_pie}
\title{Draw Pie Node}
\usage{
draw_pie(
  x,
  y,
  size,
  fill,
  border_color,
  border_width,
  alpha = 1,
  values = NULL,
  colors = NULL,
  ...
)
}
\description{
Draw a pie chart node with multiple segments.
}
\keyword{internal}
